1.8.9 Recreation (WIP 5)
Created CustomButtons for the Gui to give it the similar look of Wildfire's Female Gender Mod. it's not perfect but it's all I could do for Minecraft 1.8.9. just know I have limited components and apis to work with in this version. and I'm NOT A TOP-TIER GOD MOD DEVELOPER so I'm trying. :(

And I was successfully able to create Breasts and Physics for the Character..... But please don't look at the results ;-; https://github.com/user-attachments/assets/e0e0515e-598c-4a1f-8af7-2d1048b60b96