1.8.9 Recreation (WIP 6)
Optimized and improved the Wardrobe Gui, being "almost" faithful to the original mod. and like I said, it aims to only be a lot like the latest version of the mod. I will not include PlayerLists even thought I already created the whole code for it? (Which was only just made for debugging purposes to test the Model on other players, won't actually be implemented.)