1.8.9 Recreation (WIP 12)
IN THIS, THE GUI'S HAVE FINALLY BEEN FIXED AND UPDATED! (except the Pitch Slider)
This is the biggest WIP that mainly updates the gui and layout, options & etc.

One thing to Notice is that the Breasts Model is now on the torso but still in a work in progress.

But Next, The GUI now works exactly like Wildfire's Version, I made sure the Character Personalization GUI looks exactly like the 1.21.4+ GUI, same size, same positions and etc. ALL IN 1.8.9?!

Also this WIP 12 was also gonna integrate Cloud Sync but couldn't successfully make that a reality.
So instead, I changed the button into the Theme Button! So if you interact with the button, you can change the GUI's theme to Dark Mode! which is a major addition to WIP 12!

Fun Fact: The Theme Button doesn't exist in the original Female Gender Mod so this makes 1.8.9 Recreation standout, unless Wildfire wants it, they have full permission to take the Dark Mode textures and the ability to change the theme!